backend/requirements.txt
django
djangorestframework
stripe
psycopg2-binary
gunicorn
whitenoise
django-cors-headers
reportlab
django-anymail[sendgrid]
django-storages
boto3
Pillow
python-dotenv
